requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
google-genai
uvicorn
fastapi
pydantic
python-multipart
markdown-it-py